requirements.txt
toml
polars[all]
lxml
pillow